Move code from bazelisk.go into core package.
As a result, Bazelisk can be used as a real library, with bazelisk.go being a very thin wrapper.
