+++ BUILD
@@ -64,6 +64,7 @@ go_test(
     embed = [":go_default_library"],
     importpath = "github.com/bazelbuild/bazelisk",
     deps = [
+        "//core",
         "@io_bazel_rules_go//go/tools/bazel:go_default_library",
     ],
 )
